Refacto: Merged features library heade, filters and viewer into one feature
