+++ lib/main.dart
@@ -1,9 +1,7 @@
-import 'package:fluster_media_center/features/LibraryFilters/screens/library_filters.dart';
+import 'package:fluster_media_center/features/LibraryFilters/screens/library.dart';
 import 'package:flutter/material.dart';
 import 'package:fluster_media_center/src/rust/frb_generated.dart';
 import 'package:fluster_media_center/features/NavigatorRail/screens/navigator_rail.dart';
-import 'package:fluster_media_center/features/LibraryView/screens/library_view.dart';
-import 'package:fluster_media_center/features/LibraryHeader/screens/library_header.dart';
 import 'package:media_kit/media_kit.dart';
 
 Future<void> main() async {
@@ -30,21 +28,7 @@ class MyApp extends StatelessWidget {
         body: Row(
           children: [
             Expanded(flex: 1, child: NavigatorRail()),
-            Expanded(
-              flex: 4,
-              child: Padding(
-                padding: const EdgeInsets.all(20.0),
-                child: CustomScrollView(
-                  slivers: [
-                    LibraryHeader(),
-                    SliverToBoxAdapter(child: SizedBox(height: 20)),
-                    LibraryFilters(),
-                    SliverToBoxAdapter(child: SizedBox(height: 20)),
-                    LibraryView(),
-                  ],
-                ),
-              ),
-            ),
+            Expanded(flex: 4, child: Library()),
           ],
         ),
       ),
